推流第一版

+++ main/hal/board/config.h
@@ -42,11 +42,11 @@
 
 
 /* Camera pins */
-#define CAMERA_PIN_PWDN ((uint32_t)GPIO_NUM_NC)
-#define CAMERA_PIN_RESET ((uint32_t)GPIO_NUM_NC)
-#define CAMERA_PIN_XCLK  ((uint32_t)GPIO_NUM_NC) // 像素时钟 (固定由 20MHz 外部晶振输入) 
-#define CAMERA_PIN_SIOD ((uint32_t)GPIO_NUM_NC)  // 串行时钟 Using existing I2C port
-#define CAMERA_PIN_SIOC ((uint32_t)GPIO_NUM_NC)  // 串行时钟 Using existing I2C port
+#define CAMERA_PIN_PWDN GPIO_NUM_NC
+#define CAMERA_PIN_RESET GPIO_NUM_NC
+#define CAMERA_PIN_XCLK  GPIO_NUM_NC // 像素时钟 (固定由 20MHz 外部晶振输入) 
+#define CAMERA_PIN_SIOD GPIO_NUM_NC  // 串行时钟 Using existing I2C port
+#define CAMERA_PIN_SIOC GPIO_NUM_NC  // 串行时钟 Using existing I2C port
 #define CAMERA_PIN_D0 GPIO_NUM_39
 #define CAMERA_PIN_D1 GPIO_NUM_40
 #define CAMERA_PIN_D2 GPIO_NUM_41
